Add files via upload
I have created a dashboard using Tableau a data visualiztion tool. This dashboard is in possession of info on different airlines and variations among their prices on the basis of time of booking,location and so on. In this repository I have uploaded two files one is a tableau link containing my dashboard and other one is a docx that has info and questionaire that my dashboard will answer.